Fix warnings for 64 bit

+++ src/Lt/Platforms/Windows/Graphics/CpuWindow.cpp
@@ -20,7 +20,7 @@ void Lt::Graphics::CpuWindow::Present(Lt::u8* pixels)
     _BITMAPINFO.bmiHeader.biBitCount = 32;
     _BITMAPINFO.bmiHeader.biCompression = BI_RGB;
 
-    SetDIBitsToDevice(_Window._HDC, 0, 0, _Window.Size().PosX(), _Window.Size().PosY(), 0, 0, 0, _Window.Size().PosY(), pixels, &_BITMAPINFO, DIB_RGB_COLORS);
+    SetDIBitsToDevice(_Window._HDC, 0, 0, (DWORD)_Window.Size().PosX(), (DWORD)_Window.Size().PosY(), 0, 0, 0, (UINT)_Window.Size().PosY(), pixels, &_BITMAPINFO, DIB_RGB_COLORS);
 }
 
 const Lt::Graphics::Point2i& Lt::Graphics::CpuWindow::Size()
